requirements.txt
docker
requests
flask
kafka-python
python-logstash
jaeger-client
prometheus-client
psutil
minio